Removed Multi-Threading
Made Mesh Width and Height defined intead of just square in terms of
number of vertices. Removed Multi-threading because if the GL_Context is
defined in the main thread no image is rendered. Might re-introduce
Multi-threading to separate the user input from the rendering thread. In
order to do this, I would have to separate my initSDL() function into
two parts where OpenGL specific code (since the SDL portion is needed in
the main thread for user input) is called in the render thread.